requirement.txt
tweety-ns
pyyaml
kafka-python
pyspark
pandas
python-dotenv
apache-airflow==2.7.3 --constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.7.3/constraints-3.10.txt"